working on lights ...

+++ Viewer/include/Renderer.h
@@ -23,11 +23,11 @@ private:
 	//void BotTriangle(const glm::ivec2& maxp, const glm::ivec2& medp, const glm::ivec2& med2p, const glm::vec3& color);
 	float max_point(const float x, const float y, const float z);
 	float min_point(const float x, const float y, const float z);
-	float cal_area(const glm::ivec2& p1, const glm::ivec2& p2, const glm::ivec2& p3);
-	void DrawTriangle(const glm::ivec3& p1, const glm::ivec3& p2, const glm::ivec3& p3, const glm::vec3& color);
-	void DrawLine(const glm::ivec2& p1, const glm::ivec2& p2, const glm::vec3& color);
-	void plotLineLow(const glm::ivec2& p1, const glm::ivec2& p2, const glm::vec3& color);
-	void plotLineHigh(const glm::ivec2& p1, const glm::ivec2& p2, const glm::vec3& color);
+	float cal_area(const glm::vec2& p1, const glm::vec2& p2, const glm::vec2& p3);
+	void DrawTriangle(const glm::vec3& p1, const glm::vec3& p2, const glm::vec3& p3, const glm::vec3& color);
+	void DrawLine(const glm::vec2& p1, const glm::vec2& p2, const glm::vec3& color);
+	void plotLineLow(const glm::vec2& p1, const glm::vec2& p2, const glm::vec3& color);
+	void plotLineHigh(const glm::vec2& p1, const glm::vec2& p2, const glm::vec3& color);
 
 	void CreateBuffers(int w, int h);
 	void CreateOpenGLBuffer();
